+++ src/main/java/net/oneandone/lavender/cli/War.java
@@ -55,7 +55,7 @@ public class War extends Base {
         try (Pool pool = globals.pool()) {
             cacheroot = globals.cacheroot();
             engine = new WarEngine(cacheroot, Distributor.open(cacheroot, cluster.connect(pool), docroot, indexName),
-                    properties.modern, properties.secrets, war, outputNodesFile, nodes);
+                    properties.modern, properties.secrets, properties.bitbucketHosts, war, outputNodesFile, nodes);
             engine.run();
         }
         outputNodesFile.deleteFile();